+++ migrations/2019-03-05-233828_create_comment/up.sql
@@ -1,8 +1,8 @@
 create table comment (
-  id serial primary key,
-  creator_id int references user_ on update cascade on delete cascade not null,
-  post_id int references post on update cascade on delete cascade not null,
-  parent_id int references comment on update cascade on delete cascade,
+  id bigserial primary key,
+  creator_id bigint references user_ on update cascade on delete cascade not null,
+  post_id bigint references post on update cascade on delete cascade not null,
+  parent_id bigint references comment on update cascade on delete cascade,
   content text not null,
   removed boolean default false not null,
   read boolean default false not null,
@@ -11,19 +11,19 @@ create table comment (
 );
 
 create table comment_like (
-  id serial primary key,
-  user_id int references user_ on update cascade on delete cascade not null,
-  comment_id int references comment on update cascade on delete cascade not null,
-  post_id int references post on update cascade on delete cascade not null,
+  id bigserial primary key,
+  user_id bigint references user_ on update cascade on delete cascade not null,
+  comment_id bigint references comment on update cascade on delete cascade not null,
+  post_id bigint references post on update cascade on delete cascade not null,
   score smallint not null, -- -1, or 1 for dislike, like, no row for no opinion
   published timestamp not null default now(),
   unique(comment_id, user_id)
 );
 
 create table comment_saved (
-  id serial primary key,
-  comment_id int references comment on update cascade on delete cascade not null,
-  user_id int references user_ on update cascade on delete cascade not null,
+  id bigserial primary key,
+  comment_id bigint references comment on update cascade on delete cascade not null,
+  user_id bigint references user_ on update cascade on delete cascade not null,
   published timestamp not null default now(),
   unique(comment_id, user_id)
 );
