+++ migrations/2019-10-19-052737_create_user_mention/up.sql
@@ -1,7 +1,7 @@
 create table user_mention (
-  id serial primary key,
-  recipient_id int references user_ on update cascade on delete cascade not null,
-  comment_id int references comment on update cascade on delete cascade not null,
+  id bigserial primary key,
+  recipient_id bigint references user_ on update cascade on delete cascade not null,
+  comment_id bigint references comment on update cascade on delete cascade not null,
   read boolean default false not null,
   published timestamp not null default now(),
   unique(recipient_id, comment_id)
